update on using screen- kill

+++ kill.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 
-exec 2>/dev/null
-
-killall -9 test.sh
-killall -9 rcssserver
-killall -9 rcssserver.bin
-
-rm -f "/tmp/autotest::temp"
-
+if [ $# = 1 ]; then
+  screen -XS $1 quit
+else
+  exec 2>/dev/null
+  killall -9 test.sh
+  killall -9 rcssserver
+  killall -9 rcssserver.bin
+  rm -f "/tmp/autotest::temp"
+fi
